Improve getting toc for front matter
